added extra check on bulk uploader to avoid errors

+++ code/import/CatalogueProductCSVBulkLoader.php
@@ -29,6 +29,7 @@ class CatalogueProductCSVBulkLoader extends CsvBulkLoader
         if (class_exists("Product")) {
             if (!$objectClass || $objectClass == "CatalogueProduct") {
                 $objectClass = 'Product';
+                $this->objectClass = 'Product';
             }
         }
 
@@ -40,51 +41,54 @@ class CatalogueProductCSVBulkLoader extends CsvBulkLoader
 
         // Get Current Object
         $objID = parent::processRecord($record, $columnMap, $results, $preview);
+
         $object = DataObject::get_by_id($this->objectClass, $objID);
 
         $this->extend("onBeforeProcess", $object, $record, $columnMap, $results, $preview);
         
-        // Loop through all fields and setup associations
-        foreach ($record as $key => $value) {
+        if ($object != null) {
+            // Loop through all fields and setup associations
+            foreach ($record as $key => $value) {
 
-            // Find any categories (denoted by a 'CategoryXX' column)
-            if (strpos($key, 'Category') !== false) {
-                $category = CatalogueCategory::get()
-                    ->filter("Title", $value)
-                    ->first();
+                // Find any categories (denoted by a 'CategoryXX' column)
+                if (strpos($key, 'Category') !== false) {
+                    $category = CatalogueCategory::get()
+                        ->filter("Title", $value)
+                        ->first();
 
-                if ($category) {
-                    $object->Categories()->add($category);
+                    if ($category) {
+                        $object->Categories()->add($category);
+                    }
+                }
+                
+                // Find any Images (denoted by a 'ImageXX' column)
+                if (strpos($key, 'Image') !== false && $key != "Images") {
+                    $image = Image::get()
+                        ->filter("Name", $value)
+                        ->first();
+
+                    if ($image) {
+                        $object->Images()->add($image);
+                    }
+                }
+                
+                // Find any related products (denoted by a 'RelatedXX' column)
+                if (strpos($key, 'Related') !== false && $key != "RelatedProducts") {
+                    $product = Product::get()
+                        ->filter("StockID", $value)
+                        ->first();
+
+                    if ($product) {
+                        $object->RelatedProducts()->add($product);
+                    }
                 }
             }
-            
-            // Find any Images (denoted by a 'ImageXX' column)
-            if (strpos($key, 'Image') !== false && $key != "Images") {
-                $image = Image::get()
-                    ->filter("Name", $value)
-                    ->first();
 
-                if ($image) {
-                    $object->Images()->add($image);
-                }
-            }
-            
-            // Find any related products (denoted by a 'RelatedXX' column)
-            if (strpos($key, 'Related') !== false && $key != "RelatedProducts") {
-                $product = Product::get()
-                    ->filter("StockID", $value)
-                    ->first();
+            $this->extend("onAfterProcess", $object, $record, $columnMap, $results, $preview);
 
-                if ($product) {
-                    $object->RelatedProducts()->add($product);
-                }
-            }
+            $object->destroy();
+            unset($object);
         }
-
-        $this->extend("onAfterProcess", $object, $record, $columnMap, $results, $preview);
-
-        $object->destroy();
-        unset($object);
 
         return $objID;
     }
